Conflicts:
	backends/gdx-backend-android/src/com/badlogic/gdx/backends/android/AndroidGraphicsDaydream.java
	backends/gdx-backend-robovm/src/com/badlogic/gdx/backends/iosrobovm/IOSGraphics.java

+++ gdx/src/com/badlogic/gdx/input/GestureDetector.java
@@ -261,7 +261,7 @@ public class GestureDetector extends InputAdapter {
 		inTapSquare = false;
 	}
 
-	public void setTapSquareSize (int tapSquareSize) {
+	public void setTapSquareSize (float tapSquareSize) {
 		this.tapSquareSize = tapSquareSize;
 	}
